+++ src/main/kotlin/omok/domain/gameState/WhiteTurn.kt
@@ -2,7 +2,7 @@ package omok.domain.gameState
 
 import omok.domain.OmokBoard
 import omok.domain.OmokPoint
-import omok.domain.omokRule.adapter.RuleAdapter
+import omok.domain.omokRule.adapter.Referee
 import omok.domain.state.WhiteStoneState
 
 class WhiteTurn(override val omokBoard: OmokBoard) : GameState {
@@ -10,11 +10,11 @@ class WhiteTurn(override val omokBoard: OmokBoard) : GameState {
     override val isRunning: Boolean = true
 
     override fun play(point: OmokPoint): GameState {
-        val adapter = RuleAdapter()
+        val referee = Referee()
 
         return when {
-            adapter.isWin(omokBoard, point, stoneState) -> WhiteWin(omokBoard.placeStone(point, stoneState))
-            adapter.isForbidden(omokBoard, point) -> this
+            referee.isWin(omokBoard, point, stoneState) -> WhiteWin(omokBoard.placeStone(point, stoneState))
+            referee.isForbidden(omokBoard, point) -> this
             else -> BlackTurn(omokBoard.placeStone(point, stoneState))
         }
     }
