+++ src/main/kotlin/omok/domain/omokRule/adapter/RenjuRule.kt
@@ -1,0 +1,8 @@
+package omok.domain.omokRule.adapter
+
+import omok.domain.OmokBoard
+import omok.domain.OmokPoint
+
+interface RenjuRule {
+    fun isForbidden(omokBoard: OmokBoard, omokPoint: OmokPoint): Boolean
+}
