+++ src/test/kotlin/omok/domain/omokRule/FourRenjuRuleTest.kt
@@ -2,16 +2,16 @@ package omok.domain.omokRule
 
 import omok.domain.OmokBoard
 import omok.domain.OmokPoint
-import omok.domain.omokRule.adapter.RuleAdapter
+import omok.domain.omokRule.adapter.Referee
 import omok.domain.state.BlackStoneState
 import omok.domain.state.WhiteStoneState
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class ThreeRuleTest {
+class FourRenjuRuleTest {
     private var omokBoard = OmokBoard()
-    private var adapter = RuleAdapter()
+    private var referee = Referee()
 
     @BeforeEach
     fun setUp() {
@@ -19,33 +19,38 @@ class ThreeRuleTest {
     }
 
     /*
-        1 0 0
-        1 0 0
-        0 1 1
+        1 0 0 0
+        1 0 0 0
+        1 0 0 0
+        0 1 1 1
      */
     @Test
-    fun `열린 3이 두개 이상이면 둘 수 없다`() {
+    fun `검은 돌은 열린 4가 두개 이상이면 둘 수 없다`() {
         omokBoard = omokBoard.placeStone(OmokPoint(5, 9), BlackStoneState)
         omokBoard = omokBoard.placeStone(OmokPoint(5, 10), BlackStoneState)
+        omokBoard = omokBoard.placeStone(OmokPoint(5, 11), BlackStoneState)
         omokBoard = omokBoard.placeStone(OmokPoint(6, 8), BlackStoneState)
         omokBoard = omokBoard.placeStone(OmokPoint(7, 8), BlackStoneState)
+        omokBoard = omokBoard.placeStone(OmokPoint(8, 8), BlackStoneState)
         omokBoard = omokBoard.placeStone(OmokPoint(1, 1), WhiteStoneState)
 
-        assertThat(adapter.isForbidden(omokBoard, OmokPoint(5, 8))).isTrue
+        assertThat(referee.isForbidden(omokBoard, OmokPoint(5, 8))).isTrue
     }
 
     /*
-         1
-         1
-         0
+        1
+        1
+        1
+        0
      */
     @Test
-    fun `열린 3이 한 개면 둘 수 있다`() {
+    fun `검은 돌은 열린 4가 한 개면 둘 수 있다`() {
         omokBoard = omokBoard.placeStone(OmokPoint(5, 9), BlackStoneState)
         omokBoard = omokBoard.placeStone(OmokPoint(5, 10), BlackStoneState)
+        omokBoard = omokBoard.placeStone(OmokPoint(5, 11), BlackStoneState)
         omokBoard = omokBoard.placeStone(OmokPoint(1, 1), WhiteStoneState)
 
-        assertThat(adapter.isForbidden(omokBoard, OmokPoint(5, 8))).isFalse
+        assertThat(referee.isForbidden(omokBoard, OmokPoint(5, 8))).isFalse
     }
 
     /*
@@ -54,7 +59,7 @@ class ThreeRuleTest {
         0 1 1 1
      */
     @Test
-    fun `열린 3 4면 둘 수 있다`() {
+    fun `검은 돌은 열린 3 4면 둘 수 있다`() {
         omokBoard = omokBoard.placeStone(OmokPoint(5, 9), BlackStoneState)
         omokBoard = omokBoard.placeStone(OmokPoint(5, 10), BlackStoneState)
         omokBoard = omokBoard.placeStone(OmokPoint(6, 8), BlackStoneState)
@@ -62,6 +67,6 @@ class ThreeRuleTest {
         omokBoard = omokBoard.placeStone(OmokPoint(8, 8), BlackStoneState)
         omokBoard = omokBoard.placeStone(OmokPoint(1, 1), WhiteStoneState)
 
-        assertThat(adapter.isForbidden(omokBoard, OmokPoint(5, 8))).isFalse
+        assertThat(referee.isForbidden(omokBoard, OmokPoint(5, 8))).isFalse
     }
 }
